requirements.txt
piper-tts
onnxruntime
numpy
torch
torchaudio
librosa
soundfile
pydub
scipy
matplotlib
tensorboard
jsonlines
